Real-valued features
For some reason, previous version lacked a description. Anyway, this is a class designed with the intent of investigating 3-ways co-activations in dynamic functional connectivity matrices of fMRI data in order to discriminate between RR and healthy subjects, following the (rather strong) hypothesis that it is possible to observe a re-route through a working connection to compensate for a missing one due to degeneration, leading to the systematic appearance of 3-ways co-activitations absent (or less likely) in healthy subjects. In this specific build, the features are not binary, but rather represent the difference between 3-way co-activation for a current subject compared against the median connection strength on healthy subjects. There are a few issues to solve, still: permutation test is imperfect, classifier is suboptimal.