Elastic net failure
There was a major error in my understanding of how elastic net works, as the cross-validation matlab implements for that function has nothing to do with the cross-validation for classifiers: for elastic net, ONE model is produced following cross-validation, whereas for classifiers ONE MODEL PER FOLD is produced following cross-validation. The correct way of using that is through a nested cross-validation, but that led nowhere... Back to the drawing board!